requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
databricks-sdk==0.46.0
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
executing==2.2.0
fastapi==0.115.11
Flask==3.1.0
fonttools==4.56.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenery==3.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hyperopt==0.2.7
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonref==1.1.0
jsonschema==4.20.0
jsonschema-specifications==2024.10.1
jsonsubschema==0.0.7
kafka-python==2.0.6
kiwisolver==1.4.8
lale==0.8.4
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mlflow==2.21.0
mlflow-skinny==2.21.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-nccl-cu12==2.26.2
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
portion==2.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.4.2
scikit-surprise==1.1.4
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
zipp==3.21.0
dill==0.3.7
flask-limiter==3.12
flask-caching==2.3.1
cachelib==0.13.0
limits==5.1.0